Implement tuples and pattern matching
A pattern matches a variable if they have the same structure (tuples
of same length & same nesting), and primitives (int, bool) match values.
In this case names occurring in the pattern have the corresponding
value bound to them. Environments are now mutable (bind() method)
to allow this.
Patterns can appear on the LHS of the equals sign in `let` expressions,
and on the parameters of functions.